Config services splitted into different files
