+++ packages/apps/job-launcher/server/src/app.module.ts
@@ -12,7 +12,7 @@ import { UserModule } from './modules/user/user.module';
 import { JobModule } from './modules/job/job.module';
 import { PaymentModule } from './modules/payment/payment.module';
 import { Web3Module } from './modules/web3/web3.module';
-import { envValidator } from './common/config';
+import { envValidator } from './common/config/env-schema';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 import { StorageModule } from './modules/storage/storage.module';
