+++ packages/apps/job-launcher/server/src/common/config/config.module.ts
@@ -1,16 +1,15 @@
 import { Module, Global } from '@nestjs/common';
-import {
-  CommonConfigService,
-  AuthConfigService,
-  DatabaseConfigService,
-  Web3ConfigService,
-  S3ConfigService,
-  StripeConfigService,
-  SendgridConfigService,
-  CvatConfigService,
-  PGPConfigService,
-} from './env';
 import { ConfigService } from '@nestjs/config';
+
+import { AuthConfigService } from './auth-config.service';
+import { CommonConfigService } from './common-config.service';
+import { CvatConfigService } from './cvat-config.service';
+import { DatabaseConfigService } from './database-config.service';
+import { PGPConfigService } from './pgp-config.service';
+import { S3ConfigService } from './s3-config.service';
+import { SendgridConfigService } from './sendgrid-config.service';
+import { StripeConfigService } from './stripe-config.service';
+import { Web3ConfigService } from './web3-config.service';
 
 @Global()
 @Module({
