+++ packages/apps/job-launcher/server/src/common/config/pgp-config.service.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+
+@Injectable()
+export class PGPConfigService {
+  constructor(private configService: ConfigService) {}
+  get encrypt(): boolean {
+    return this.configService.get<string>('PGP_ENCRYPT', 'false') === 'true';
+  }
+  get privateKey(): string {
+    return this.configService.get<string>('PGP_PRIVATE_KEY', '');
+  }
+  get passphrase(): string {
+    return this.configService.get<string>('PGP_PASSPHRASE', '');
+  }
+}
