+++ packages/apps/job-launcher/server/src/common/config/sendgrid-config.service.ts
@@ -1,0 +1,22 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+
+@Injectable()
+export class SendgridConfigService {
+  constructor(private configService: ConfigService) {}
+  get apiKey(): string {
+    return this.configService.get<string>('SENDGRID_API_KEY', '');
+  }
+  get fromEmail(): string {
+    return this.configService.get<string>(
+      'SENDGRID_FROM_EMAIL',
+      'job-launcher@hmt.ai',
+    );
+  }
+  get fromName(): string {
+    return this.configService.get<string>(
+      'SENDGRID_FROM_NAME',
+      'Human Protocol Job Launcher',
+    );
+  }
+}
