+++ packages/apps/job-launcher/server/src/modules/storage/storage.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 import { StorageService } from './storage.service';
-import { ConfigModule } from '@nestjs/config';
-import { s3Config } from '../../common/config';
 import { EncryptionModule } from '../encryption/encryption.module';
 
 @Module({
-  imports: [ConfigModule.forFeature(s3Config), EncryptionModule],
+  imports: [EncryptionModule],
   providers: [StorageService],
   exports: [StorageService],
 })
