+++ packages/apps/job-launcher/server/src/modules/storage/storage.service.ts
@@ -5,35 +5,32 @@ import {
 } from '@human-protocol/sdk';
 import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import * as Minio from 'minio';
-import { S3ConfigType, s3ConfigKey } from '../../common/config';
-import { ErrorBucket } from '../../common/constants/errors';
 import stringify from 'json-stable-stringify';
+import { ErrorBucket } from '../../common/constants/errors';
 import { ContentType, Extension } from '../../common/enums/storage';
 import { UploadedFile } from '../../common/interfaces';
-import { ConfigService } from '@nestjs/config';
+import { S3ConfigService } from '../../common/config/s3-config.service';
 
 @Injectable()
 export class StorageService {
   public readonly minioClient: Minio.Client;
 
   constructor(
-    @Inject(s3ConfigKey)
-    private s3Config: S3ConfigType,
-    public readonly configService: ConfigService,
+    public readonly s3ConfigService: S3ConfigService,
     @Inject(Encryption) private readonly encryption: Encryption,
   ) {
     this.minioClient = new Minio.Client({
-      endPoint: this.s3Config.endPoint,
-      port: this.s3Config.port,
-      accessKey: this.s3Config.accessKey,
-      secretKey: this.s3Config.secretKey,
-      useSSL: this.s3Config.useSSL,
+      endPoint: this.s3ConfigService.endpoint,
+      port: this.s3ConfigService.port,
+      accessKey: this.s3ConfigService.accessKey,
+      secretKey: this.s3ConfigService.secretKey,
+      useSSL: this.s3ConfigService.useSSL,
     });
   }
   public formatUrl(key: string): string {
-    return `${this.s3Config.useSSL ? 'https' : 'http'}://${
-      this.s3Config.endPoint
-    }:${this.s3Config.port}/${this.s3Config.bucket}/${key}`;
+    return `${this.s3ConfigService.useSSL ? 'https' : 'http'}://${
+      this.s3ConfigService.endpoint
+    }:${this.s3ConfigService.port}/${this.s3ConfigService.bucket}/${key}`;
   }
 
   public async download(url: string): Promise<any> {
@@ -56,7 +53,7 @@ export class StorageService {
     data: string | object,
     hash: string,
   ): Promise<UploadedFile> {
-    if (!(await this.minioClient.bucketExists(this.s3Config.bucket))) {
+    if (!(await this.minioClient.bucketExists(this.s3ConfigService.bucket))) {
       throw new BadRequestException(ErrorBucket.NotExist);
     }
 
@@ -68,10 +65,15 @@ export class StorageService {
     const key = `s3${hash}${isStringData ? '' : Extension.JSON}`;
 
     try {
-      await this.minioClient.putObject(this.s3Config.bucket, key, content, {
-        'Content-Type': contentType,
-        'Cache-Control': 'no-store',
-      });
+      await this.minioClient.putObject(
+        this.s3ConfigService.bucket,
+        key,
+        content,
+        {
+          'Content-Type': contentType,
+          'Cache-Control': 'no-store',
+        },
+      );
 
       return { url: this.formatUrl(key), hash };
     } catch (e) {
